backend/requirements.txt
# Зависимости Python для backend
# Например:
# fastapi==0.95.0
# pydicom==2.4.0
# simpleitk==2.2.0

fastapi
pydicom
numpy
pillow
trimesh
meshio
matplotlib
